requirements.txt
nltk==3.8.1
numpy==1.24.4
pandas==2.0.3
pcfg==0.1.5
scikit_learn==1.3.0
sklearn_crfsuite==0.3.6
spacy==3.6.0
spacy_conll==3.4.0
spacy_stanza==1.0.3
stanza==1.5.0
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
wget==3.2
